Groundwork for supporting multiple releases
Related to https://github.com/dhall-lang/dhall-kubernetes/issues/75

This generalizes the machinery for fetching the Kubernetes OpenAPI spec to
support fetching multiple versions.  In particular, the Kubernetes version
is no longer tied to the version of Nixpkgs that we use.

We still only build generate Dhall types for version 1.16, but in a
subsequent change we can start mass-generating support for multiple
Dhall packages (one per Kubernetes version)